Add Ability to Get CampaignMessage from Campaign
As a part of addressing iFixit/ifixit#33568, add the ability to retrieve
CampaignMessage objects when provided a campaignid. Do this by adding
the getCampaignMessage function. This allows us to retrieve a campaign's
campaign message id, which we need in order to properly retrieve a
campaign's message.

Also, edit getMessage to reflect that it should take in a campaign
message id (not a campaign id) and that the purpose of the function is
to retrieve a Message object, not a CampaignMessage object.